Метод create добавил его работу через ResponseEntity - все создает, и ошибку валидации прокидывает

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -1,9 +1,13 @@
 package ru.kata.spring.boot_security.demo.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,8 +18,10 @@ import org.springframework.web.bind.annotation.RestController;
 import ru.kata.spring.boot_security.demo.entity.Role;
 import ru.kata.spring.boot_security.demo.entity.User;
 import ru.kata.spring.boot_security.demo.service.UserService;
+import ru.kata.spring.boot_security.demo.utill.UserNotCreatedException;
 import ru.kata.spring.boot_security.demo.utill.UserNotFoundException;
 
+import javax.validation.Valid;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -48,11 +54,20 @@ public class AdminController {
 
     //	Добавить Юзера
     @PostMapping("/create")
-    public User create(@RequestBody User user, @RequestParam("role") String roleName) {
+    public ResponseEntity<HttpStatus> create(@RequestBody @Valid User user, BindingResult bindingResult, @RequestParam("role") String roleName) {
+        if (bindingResult.hasErrors()) {
+            StringBuilder errorsMsg = new StringBuilder();
+
+            List<FieldError> fieldErrors = bindingResult.getFieldErrors();
+            for (FieldError fieldError : fieldErrors) {
+                errorsMsg.append(fieldError.getField()).append(" :!!!!!!!!! ").append(fieldError.getDefaultMessage()).append("\n");
+            }
+            throw new UserNotCreatedException(errorsMsg.toString());
+        }
         Role role = userService.findRoleByName(roleName);
         user.setRoles(Set.of(role));// Устанавливаем новую роль
         userService.save(user);
-        return user;
+        return ResponseEntity.ok(HttpStatus.OK);
     }
 
     //  Изменить Юзера
@@ -93,6 +108,9 @@ public class AdminController {
 
     @DeleteMapping("/delete")
     public String deleteUser(@RequestParam int id) {
+        if (userService.getById(id) == null) {
+            throw new UserNotFoundException("User с таким ID = " + id + " в базе-данных нет");
+        }
         userService.delete(id);
         return "User with ID = " + id + " deleted";
     }
